Dak Details With Custom Date Picker

+++ src/dNothi.Core/Entities/SyncStatus.cs
@@ -9,6 +9,6 @@ namespace dNothi.Core.Entities
 {
     public class SyncStatus : BaseEntity
     {
-        public int dak_id { get; set; }
+        public long dak_id { get; set; }
     }
 }
